[2525]오븐시계
잘 풀어보았습니당
